--- app/src/main/java/com/Equipo20/vinilos_front_mobile_mvtqn/models/Album.kt
@@ -1,11 +1,0 @@
-package com.Equipo20.vinilos_front_mobile_mvtqn.models
-
-data class Album (
-    val albumId:Int,
-    val name:String,
-    val cover:String,
-    val releaseDate:String,
-    val description:String,
-    val genre:String,
-    val recordLabel:String
-)
