--- app/src/main/java/com/Equipo20/vinilos_front_mobile_mvtqn/models/Artist.kt
@@ -1,15 +1,0 @@
-package com.Equipo20.vinilos_front_mobile_mvtqn.models
-
-class Artist (
-
-    val artistId: Int,
-    val artistName:String,
-    val artistImage:String,
-    val artisDescription:String
-
-              )
-{
-
-
-
-}
