--- app/src/main/java/com/Equipo20/vinilos_front_mobile_mvtqn/models/Collector.kt
@@ -1,8 +1,0 @@
-package com.Equipo20.vinilos_front_mobile_mvtqn.models
-
-data class Collector (
-    val collectorId: Int,
-    val name:String,
-    val telephone:String,
-    val email:String
-)
