--- app/src/main/java/com/Equipo20/vinilos_front_mobile_mvtqn/ui/artists/ArtistsViewModel.kt
@@ -1,13 +1,0 @@
-package com.Equipo20.vinilos_front_mobile_mvtqn.ui.artists
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class ArtistsViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is artists Fragment"
-    }
-    val text: LiveData<String> = _text
-}
