--- app/src/main/java/com/Equipo20/vinilos_front_mobile_mvtqn/ui/collectors/CollectorsViewModel.kt
@@ -1,13 +1,0 @@
-package com.Equipo20.vinilos_front_mobile_mvtqn.ui.collectors
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class CollectorsViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is collectors Fragment"
-    }
-    val text: LiveData<String> = _text
-}
