--- app/src/main/java/com/Equipo20/vinilos_front_mobile_mvtqn/viewmodels/ArtistViewModel.kt
@@ -1,54 +1,0 @@
-package com.Equipo20.vinilos_front_mobile_mvtqn.viewmodels
-
-import android.app.Application
-import androidx.lifecycle.*
-import com.Equipo20.vinilos_front_mobile_mvtqn.models.Artist
-import com.Equipo20.vinilos_front_mobile_mvtqn.network.NetworkServiceAdapter
-
-class ArtistViewModel(application:Application): AndroidViewModel(application)  {
-
-    private val _artist = MutableLiveData<List<Artist>>()
-
-    val artist: LiveData<List<Artist>>
-        get() = _artist
-
-    private var _eventNetworkError = MutableLiveData<Boolean>(false)
-
-    val eventNetworkError: LiveData<Boolean>
-        get() = _eventNetworkError
-
-    private var _isNetworkErrorShown = MutableLiveData<Boolean>(false)
-
-    val isNetworkErrorShown: LiveData<Boolean>
-        get() = _isNetworkErrorShown
-
-    init {
-        refreshDataFromNetwork()
-    }
-
-    private fun refreshDataFromNetwork() {
-        NetworkServiceAdapter.getInstance(getApplication()).getArtist({
-            _artist.postValue(it)
-            _eventNetworkError.value = false
-            _isNetworkErrorShown.value = false
-        },{
-            _eventNetworkError.value = true
-        })
-    }
-
-    fun onNetworkErrorShown() {
-        _isNetworkErrorShown.value = true
-    }
-
-    class Factory(val app: Application) : ViewModelProvider.Factory {
-        override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            if (modelClass.isAssignableFrom(ArtistViewModel::class.java)) {
-                @Suppress("UNCHECKED_CAST")
-                return ArtistViewModel(app) as T
-            }
-            throw IllegalArgumentException("Unable to construct viewmodel")
-        }
-    }
-}
-
-
