+++ app/src/main/java/com/grupo20/vinilos/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.Equipo20.vinilos_front_mobile_mvtqn
+package com.grupo20.vinilos
 
 import android.os.Bundle
 import com.google.android.material.bottomnavigation.BottomNavigationView
@@ -7,7 +7,7 @@ import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
-import com.Equipo20.vinilos_front_mobile_mvtqn.databinding.ActivityMainBinding
+import com.grupo20.vinilos.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
 
@@ -22,11 +22,10 @@ class MainActivity : AppCompatActivity() {
         val navView: BottomNavigationView = binding.navView
 
         val navController = findNavController(R.id.nav_host_fragment_activity_main)
-        // Passing each menu ID as a set of Ids because each
-        // menu should be considered as top level destinations.
+
         val appBarConfiguration = AppBarConfiguration(
             setOf(
-                R.id.navigation_home, R.id.navigation_artists, R.id.navigation_collectors
+                R.id.navigation_home, R.id.navigation_collectors
             )
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
