+++ app/src/main/java/com/grupo20/vinilos/ui/collectors/CollectorsViewModel.kt
@@ -1,0 +1,7 @@
+package com.grupo20.vinilos.ui.collectors
+
+import androidx.lifecycle.ViewModel
+
+class CollectorsViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
