+++ app/src/main/java/com/grupo20/vinilos/ui/home/HomeFragment.kt
@@ -1,4 +1,4 @@
-package com.Equipo20.vinilos_front_mobile_mvtqn.ui.artists
+package com.grupo20.vinilos.ui.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.Equipo20.vinilos_front_mobile_mvtqn.databinding.FragmentArtistsBinding
+import com.grupo20.vinilos.databinding.FragmentHomeBinding
 
-class ArtistsFragment : Fragment() {
+class HomeFragment : Fragment() {
 
-    private var _binding: FragmentArtistsBinding? = null
+    private var _binding: FragmentHomeBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class ArtistsFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val artistsViewModel =
-            ViewModelProvider(this).get(ArtistsViewModel::class.java)
+        val homeViewModel =
+            ViewModelProvider(this).get(HomeViewModel::class.java)
 
-        _binding = FragmentArtistsBinding.inflate(inflater, container, false)
+        _binding = FragmentHomeBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textArtists
-        artistsViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textHome
+        homeViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
