+++ app/src/main/java/com/grupo20/vinilos/ui/home/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.Equipo20.vinilos_front_mobile_mvtqn.ui.home
+package com.grupo20.vinilos.ui.home
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -7,7 +7,7 @@ import androidx.lifecycle.ViewModel
 class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "Vinilos"
     }
     val text: LiveData<String> = _text
 }
