+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -7,13 +7,8 @@
         android:title="@string/title_home" />
 
     <item
-        android:id="@+id/navigation_artists"
-        android:icon="@drawable/ic_artists_black_24dp"
-        android:title="@string/title_artists" />
-
-    <item
         android:id="@+id/navigation_collectors"
-        android:icon="@drawable/ic_collectors_black_24dp"
+        android:icon="@drawable/ic_collectors"
         android:title="@string/title_collectors" />
 
 </menu>
