+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,7 @@
 <resources>
-    <string name="app_name">Vinilos_Front_Mobile_MVTQN</string>
+    <string name="app_name">Vinilos</string>
     <string name="title_home">Home</string>
-    <string name="title_artists">Artists</string>
+    <string name="title_dashboard">Dashboard</string>
+    <string name="title_notifications">Notifications</string>
     <string name="title_collectors">Collectors</string>
 </resources>
