+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,4 @@
-#Wed Nov 02 11:09:25 COT 2022
+#Sat Nov 05 12:51:02 COT 2022
 distributionBase=GRADLE_USER_HOME
 distributionUrl=https\://services.gradle.org/distributions/gradle-7.4-bin.zip
 distributionPath=wrapper/dists
