+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "Vinilos_Front_Mobile_MVTQN"
+rootProject.name = "Vinilos"
 include ':app'
